Added validation properties to PagedRequestModel
